+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Changelog
 
+### [00010] - [18/FEB/2025] @Julio Villagrana
+- Add upload files to azure blob @azure/storage-blob
+
 ### [00009] - [31/ENE/2025] @Julio Villagrana
 - Catch web hook Labs Mobile logs error